+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -59,7 +59,7 @@ public class StatsServiceTest {
     public void calculateAbove() {
         StatsService service = new StatsService();
         long[] sale = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        long expected = 5;
+        int expected = 5;
         double actual = service.calculateAbove(sale);
         Assertions.assertEquals(expected, actual);
     }
